ZombieHitEvent renamed to ZombieTakeDmgEvent
