+++ pvzclass/Events/Events.h
@@ -15,7 +15,7 @@
 #include "NewGameEvent.h"
 #include "PlantCreateEvent.h"
 #include "PlantReloadEvent.h"
-#include "PlantRemoveEvent.h"
+#include "PlantDieEvent.h"
 #include "PlantShootEvent.h"
 #include "PeaOnFireEvent.h"
 #include "ProjectileCollisionEvent.h"
@@ -32,7 +32,7 @@
 #include "ZombieDecelerateEvent.h"
 #include "ZombieEatEvent.h"
 #include "ZombieFrozeEvent.h"
-#include "ZombieHitEvent.h"
+#include "ZombieTakeDmgEvent.h"
 #include "ZombieHypnotizeEvent.h"
 #include "ZombieRemoveEvent.h"
 #include "ZombieUpdateActionEvent.hpp"
