+++ pvzclass/pvzclass.vcxproj.filters
@@ -530,7 +530,7 @@
     <ClInclude Include="Events\ZombieFrozeEvent.h">
       <Filter>头文件\Events</Filter>
     </ClInclude>
-    <ClInclude Include="Events\ZombieHitEvent.h">
+    <ClInclude Include="Events\ZombieTakeDmgEvent.h">
       <Filter>头文件\Events</Filter>
     </ClInclude>
     <ClInclude Include="Events\ZombieHypnotizeEvent.h">
